fix: package명 변경 domain->entity
