--- src/main/java/com/oseak/myFestaBackend/domain/Member.java
@@ -1,60 +1,0 @@
-package com.oseak.myFestaBackend.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor
-@Entity
-@Table(name = "member")
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
-    private Long id;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column(nullable = false)
-    private String nickname;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private Provider provider;
-
-    private String profile;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Column(name = "is_withdrawn", nullable = false)
-    @ColumnDefault("false")
-    private boolean isWithdrawn;
-
-    @Column(name = "withdrawn_at")
-    private LocalDateTime withdrawnAt;
-
-    @Builder
-    public Member(String email, String nickname, Provider provider, String profile) {
-        this.email = email;
-        this.nickname = nickname;
-        this.provider = provider;
-        this.profile = profile;
-    }
-
-    public enum Provider {
-        local, kakao
-    }
-}
