--- src/main/java/com/oseak/myFestaBackend/domain/MemberOauthToken.java
@@ -1,48 +1,0 @@
-package com.oseak.myFestaBackend.domain;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "member_oauth_token")
-public class MemberOauthToken {
-
-    @Id
-    @Column(name = "member_id")
-    private Long memberId;
-
-    @Column(name = "access_token", nullable = false)
-    private String accessToken;
-
-    @Column(name = "refresh_token", nullable = false)
-    private String refreshToken;
-
-    @Column(name = "expires_at", nullable = false)
-    private LocalDateTime expiresAt;
-
-    @Column(name = "refresh_token_expires_at", nullable = false)
-    private LocalDateTime refreshTokenExpiresAt;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
-
-    @UpdateTimestamp
-    @Column(name = "updated_at")
-    private LocalDateTime updatedAt;
-
-    @Builder
-    public MemberOauthToken(Long memberId, String accessToken, String refreshToken, LocalDateTime expiresAt, LocalDateTime refreshTokenExpiresAt) {
-        this.memberId = memberId;
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-        this.expiresAt = expiresAt;
-        this.refreshTokenExpiresAt = refreshTokenExpiresAt;
-    }
-}
