+++ src/main/java/com/oseak/myFestaBackend/entity/MemberOauthToken.java
@@ -1,0 +1,56 @@
+package com.oseak.myFestaBackend.entity;
+
+import java.time.LocalDateTime;
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "member_oauth_token")
+public class MemberOauthToken {
+
+	@Id
+	@Column(name = "member_id")
+	private Long memberId;
+
+	@Column(name = "access_token", nullable = false)
+	private String accessToken;
+
+	@Column(name = "refresh_token", nullable = false)
+	private String refreshToken;
+
+	@Column(name = "expires_at", nullable = false)
+	private LocalDateTime expiresAt;
+
+	@Column(name = "refresh_token_expires_at", nullable = false)
+	private LocalDateTime refreshTokenExpiresAt;
+
+	@CreationTimestamp
+	@Column(name = "created_at", updatable = false)
+	private LocalDateTime createdAt;
+
+	@UpdateTimestamp
+	@Column(name = "updated_at")
+	private LocalDateTime updatedAt;
+
+	@Builder
+	public MemberOauthToken(Long memberId, String accessToken, String refreshToken, LocalDateTime expiresAt,
+		LocalDateTime refreshTokenExpiresAt) {
+		this.memberId = memberId;
+		this.accessToken = accessToken;
+		this.refreshToken = refreshToken;
+		this.expiresAt = expiresAt;
+		this.refreshTokenExpiresAt = refreshTokenExpiresAt;
+	}
+}
