+++ src/main/java/com/oseak/myFestaBackend/repository/MemberOauthTokenRepository.java
@@ -1,7 +1,8 @@
 package com.oseak.myFestaBackend.repository;
 
-import com.oseak.myFestaBackend.domain.MemberOauthToken;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.oseak.myFestaBackend.entity.MemberOauthToken;
 
 public interface MemberOauthTokenRepository extends JpaRepository<MemberOauthToken, Long> {
 }
