+++ src/main/java/com/oseak/myFestaBackend/repository/MemberRepository.java
@@ -1,10 +1,11 @@
 package com.oseak.myFestaBackend.repository;
 
-import com.oseak.myFestaBackend.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.oseak.myFestaBackend.entity.Member;
+
 public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String email);
+	Optional<Member> findByEmail(String email);
 }
